Split long CSV tests into shorter chunks (TRIVIAL)
