+++ testing/testing.mk
@@ -177,6 +177,9 @@ $(eval $(call mldb_unit_test,MLDB-761-sub-queries.py))
 $(eval $(call mldb_unit_test,MLDB-565-classifier-details.js))
 $(eval $(call mldb_unit_test,MLDB-1597-regression.py))
 $(eval $(call mldb_unit_test,MLDB-749-csv-dataset.js))
+$(eval $(call mldb_unit_test,MLDB-749-count-with-offset-limit.js))
+$(eval $(call mldb_unit_test,MLDB-749-broken-csv.js))
+$(eval $(call mldb_unit_test,MLDB-749-csv-dataset-large-file-1.js))
 $(eval $(call mldb_unit_test,MLDB-702-row-aggregators.py))
 $(eval $(call mldb_unit_test,MLDB-770-group-by-parsing.js))
 $(eval $(call mldb_unit_test,MLDB-768-order-by-missing-function.js))
